refactor: move purchaseCombine to its own file
